Added more logic to filter, created FilterDialog.tsx
