+++ Eplant/views/InteractionsViewer/Topbar.tsx
@@ -4,15 +4,15 @@ import { Unstable_Popup as Popup } from '@mui/base/Unstable_Popup'
 import { Filter, FilterAlt, QuestionMark } from '@mui/icons-material'
 import Add from '@mui/icons-material/Add'
 import Remove from '@mui/icons-material/Remove'
+import { ListItemText, ListSubheader, Switch } from '@mui/material'
 import AppBar from '@mui/material/AppBar'
 import Box from '@mui/material/Box'
 import Button from '@mui/material/Button'
 import ButtonGroup from '@mui/material/ButtonGroup'
-import Toolbar from '@mui/material/Toolbar'
-import Typography from '@mui/material/Typography'
 import List from '@mui/material/List'
 import ListItem from '@mui/material/ListItem'
-import { ListItemText, ListSubheader, Switch } from '@mui/material'
+import Toolbar from '@mui/material/Toolbar'
+import Typography from '@mui/material/Typography'
 
 interface TopbarProps {
   gene: string
@@ -23,9 +23,24 @@ const Topbar: FC<TopbarProps> = ({ gene }) => {
   const legendRef = useRef(null)
   const [showFilter, setShowFilter] = useState<boolean>(false)
   const filterRef = useRef(null)
-
+  const selectors = {
+    EPPISelector: '[type = "PPI"][method = "E"]', // checkbox
+    corrSelector: '[correlation <= ', // eppi-correlation spinner + checkbox
+    PPPISelector: '[type = "PPI"][method = "P"]', // checkbox
+    interConfSelector: '[interolog_conf <=', // PPPI correlation
+    EPDISelector: '[type = "PDI"][method = "E"]',
+    PPDISelector: '[type = "PDI"][method = "P"]',
+    fimoConfSelector: '[fimo_conf >= '
+}
   const zoomValue = 1000
 
+  const applyFilters = () =>{
+    // Create selectors
+			const eppiCorr = selectors.EPPISelector + selectors.corrSelector + EPPICorrValue + ']';
+			const pppiCorr = selectors.PPPISelector + selectors.corrSelector + PPPICorrValue + ']';
+			const pppiConf = selectors.PPPISelector + selectors.interConfSelector + PPPIConfValue + ']';
+			const ppdiConf = selectors.PPDISelector + selectors.fimoConfSelector + PPDIConfValue + ']';
+  }
   const handleLegendClick = () => {
     setShowLegend(!showLegend)
   }
